Add files via upload
Day 1  and 2. 
I have learnt what variables are, how to use the printf function. How use a comment, doubles and floats to print mathematical operations and values.
When using floats it is when you want to see decimal points for numbers that are in square roots or are long
When using a double you want a whole number using the pow function or if you want single digits.
If I want to round up I use the ceiling function and floor for rounding down.